chore(examples): Add the Poetry example
This includes a workflow that will publish a hello-world package to
ghcr.io/allexveldman/.

The version published includes the current commit sha and the name of
the example, this is to prevent collision with previous versions as well
as other (future) examples that publish the same package name.

Note that reruns of the same workflow can fail due to the version
already existing.