feat(logging): Configure logging from RUST_LOG
[EnvFilter](https://docs.rs/tracing-subscriber/latest/tracing_subscriber/filter/struct.EnvFilter.html)
takes a RUST_LOG string and creates a Filter for use in tracing.

Because cloudflare workers don't have process environment variables but
instead get the environment passed in as a parameter to fetch, logging
is initialized on the first request the worker handles.
A OnceLock is used to ensure init is only run once.