fix: Return 404 on invalid package name
Renovate will try the JSON API to get some package metadata.
this resolves to <>/package/json, which is picked up by the
`/:registry/:namespace/:package/:filename` route.

This route would return a BAD_REQUEST, since for us it is.
While we don't support the json API, instead we can return a 404 on
invalid filenames triggering the fallback in Renovate to the "simple" API.