chore(otlp): Remove use of async_channel
The log records were buffered twice in the OtlpLogLayer, once in
self.records and again in self.sender at the end of a trace.

This PR removes the async channel in favor of just the records buffer.
OtlpLogLayer now is Clone so we can keep a reference to explicitly flush
at the end of a request on the layer itself instead of having to pass
around the Receiver.