chore(otlp): Set span-kind and span attributes
This sets the span-kind for the "fetch" trace to "server" and the "send"
trace to "client" to make the spans more inline with their intent.

"method" and "path" are now included as span attributes.

This drops the "elapsed_ms" label in the log records, this
information is available as part of the traces.