chore: Adapt use of Digest
oci-spec now requires a valid digest in the DescriptorBuilder.
This updates all uses to be a Descriptor instance.

For unittests this is a bit annoying as I can no longer identify what
resource was requested based on the digest. I included comments where
applicable to help future debugging if something breaks there.