feat: Allow setting labels on the OCI image
The OCI image spec allows for "annotations" (aka labels in docker
terms) to be added to the image index and manifests.

Python allows adding "classifiers" to python package metadata.

If a packages defines a classifier in the format `PyOCI :: Label :: Key ::
Value`, a label `Key=Value` will be added to the created image.