fix: Package version list incomplete (#167)
The response to listing tags in the OCI registry can contain a `Link`
header, indicating there are more tags available. These additional tags
are now also collected when listing the versions of a package.

In addition, there is a limit to the number of versions PyOCI will fetch
the filenames for. This limit is now configurable through environment
variable `PYOCI_MAX_VERSIONS`.

To completely bypass the limit, set `PYOCI_MAX_VERSIONS=0`.
This may however cause slow responses when there is a significant number
of versions for a package.

closes: #166