feat: Include `project_urls` in  `<package>/json` (#192)
feat: Include `project_urls` in the `<package>/json`

Published packages include Package-URL entries in their request.
These URLs are used by Renovate to link to the source repo and fetch
changelogs if on a supported platform.

With this PR PyOCI will store the package URLs as the
`com.pyoci.project_urls` annotation on the
manifest in the ImageIndex.

The project_urls of the first manifest of the highest version will be
included in the `<package>/json` response.