feat(release): Allow building for multiple architectures (#266)
* feat(release): Allow building for multiple architectures

The dockerfile requires some platform-dependent parts to be able to
build with muslc so we can provide a `FROM scratch` final image.

This PR adds `docker/targer-platform.sh` to translate between buildkit's
$TARGETPLATFORM value and the platform triplet we need for the rust
build.

Currently the image is build for `linux/amd64` and `linux/arm64`.

closes: #250, related: #169

* See if we can build the ARM image on an ARM runner (#267)

This should speed up the ARM build significantly

* Smoke-test both builds of the image instead of as a separate job