feat: Support OCI sub-namespaces
The OCI distribution specification allows for multiple "/" in the namespace of a
repository.
[ref](https://github.com/opencontainers/distribution-spec/blob/main/spec.md#pulling-manifests)

PyOCI only supports a single-part namespace.
Supporting multiple parts in the namespace allows clear separation of
python packages published with pyoci from regular images in the
repository.

This PR adds the support for these extra parts in the namespace.
To not affect the way routing works, I opted for parsing and rewriting
the URI path with "/" url-encoded as "%2F" in the namespace part of the
URI before hitting the router by wrapping the Router by an
`EncodeNamespace` Service.

closes: #293