fix(auth): 502 when authenticating with `registry-1.docker.io` (#321)
`registry-1.docker.io`'s authentication response contains both a `token`
and an `access_token` which causes a 502 in pyoci.

According to the spec:
> For compatibility with OAuth 2.0, we will also accept token under the name access_token.
At least one of these fields must be specified, but both may also appear (for compatibility with older clients).
When both are specified, they should be equivalent; if they differ the client's choice is undefined.

Now we accept both fields in the response and prioritize the `token`
key.