fix: package name not consistent in list results (#338)
When listing a package containing a `-`, an URL was returned where the package name contains `_` instead.
Now the package name correctly retains the name as provided in the request.

When uploading a package, the name is now extracted from the form fields instead of parsed out of the filename.
For backwards compatibility the package is still stored in the OCI registry using the name with `_`.

related: https://github.com/AllexVeldman/pyoci/issues/329#issuecomment-3631607126