fix: Return list of package files in order
When listing a package, the results were returned unordered based on
what subrequest returned first.

As we await all results before returning, and no significant work is
done on the results, there is no reason to return this unordered.

Now the list if returned in version order, latest versions first.